# core, eth: improve delivery speed on header requests