--- a/core/headerchain.go
+++ b/core/headerchain.go
@@ -505,7 +505,7 @@ func (hc *HeaderChain) GetHeadersFrom(number, count uint64) []rlp.RawValue {
 	// If the request is for future headers, we still return the portion of
 	// headers that we are able to serve
 	if current := hc.CurrentHeader().Number.Uint64(); current < number {
-		if count >= number-current {
+		if count > number-current {
 			count -= number - current
 			number = current
 		} else {
@@ -513,7 +513,7 @@ func (hc *HeaderChain) GetHeadersFrom(number, count uint64) []rlp.RawValue {
 		}
 	}
 	var headers []rlp.RawValue
-	//If we have some of the headers in cache already, use that before going to db.
+	// If we have some of the headers in cache already, use that before going to db.
 	hash := rawdb.ReadCanonicalHash(hc.chainDb, number)
 	if hash == (common.Hash{}) {
 		return nil