--- a/core/rawdb/accessors_chain.go
+++ b/core/rawdb/accessors_chain.go
@@ -282,6 +282,12 @@ func WriteFastTxLookupLimit(db ethdb.KeyValueWriter, number uint64) {
 // ReadHeadersRLP returns the rlp-encoded headers, starting at 'number', and going
 // backwards towards genesis. This method assumes that the caller already has
 // placed a cap on count, to prevent DoS issues.
+// Since this method operates in head-towards-genesis mode, it will return an empty
+// slice in case the head ('number') is missing. Hence, the caller must ensure that
+// the head ('number') argument is actually an existing header.
+//
+// N.B: Since the input is a number, as opposed to a hash, it's implicit that
+// this method only operates on canon headers.
 func ReadHeadersRLP(db ethdb.Reader, number uint64, count uint64) []rlp.RawValue {
 	var rlpHeaders []rlp.RawValue
 	if count == 0 {
@@ -308,14 +314,17 @@ func ReadHeadersRLP(db ethdb.Reader, number uint64, count uint64) []rlp.RawValue
 			count--
 		}
 	}
-	// Then read ancients
-	for ; count > 0; i-- {
-		if data, err := db.Ancient(freezerHeaderTable, i); err == nil {
-			rlpHeaders = append(rlpHeaders, data)
-		} else {
-			break
+	if count == 0 {
+		return rlpHeaders
+	}
+	// read remaining from ancients
+	max := uint64(count * 1024)
+	data, err := db.ReadAncients(freezerHeaderTable, uint64(i+1-count), count, max)
+	if err == nil && uint64(len(data)) == count {
+		// the data is on the order [h, h+1, .., n] -- reordering needed
+		for i := range data {
+			rlpHeaders = append(rlpHeaders, data[len(data)-1-i])
 		}
-		count--
 	}
 	return rlpHeaders
 }