--- a/core/rawdb/accessors_chain.go
+++ b/core/rawdb/accessors_chain.go
@@ -297,13 +297,11 @@ func ReadHeadersRLP(db ethdb.Reader, number uint64, count uint64) []rlp.RawValue
 	if i >= limit {
 		// If we need to read live blocks, we need to figure out the hash first
 		hash := ReadCanonicalHash(db, number)
-		var hdr types.Header
 		for ; i >= limit && count > 0; i-- {
 			if data, _ := db.Get(headerKey(i, hash)); len(data) > 0 {
 				rlpHeaders = append(rlpHeaders, data)
 				// Get the parent hash for next query
-				rlp.DecodeBytes(data, &hdr)
-				hash = hdr.ParentHash
+				hash = types.HeaderParentHashFromRLP(data)
 			} else {
 				break // Maybe got moved to ancients
 			}