@holiman 2021-06-24
> This may actually make the performance a bit worse, in the case that a peer is 3-4 blocks behind us, and wants to catch up: 

Pushed a commit to use the headerchain cache better.