@holiman 2021-07-06
TODO (not in this PR), add basefee check to sanityCheck