@holiman 2021-08-13
Rebased on top of the eth/66 changes that got merged. This still does not make use of the faster ancient accessors though