@holiman 2021-09-08
Rebased again. Also, tests are added to check the correctness of the split leveldb/ancient loader