@holiman 2021-11-30
I tested this a bit on a live node. On my node, I made it 

- When the 'new' code path was requested, it did both a old-style lookup and a new-style lookup
- When doing so, it randomized: sometimes doing old style first, sometimes doing new-style first, 
- After pulling both answers, it compared the byte output of the two, to see if they matched. 

Example: 
```
INFO [11-30|11:28:29.065] Served block headers                     newtime=1.877586ms  oldtime=8.107161ms  order="new version first" reverse=false skip=0 amount=192 origin="{Hash:0x0000000000000000000000000000000000000000000000000000000000000000 Number:8166928}"
```
I then collected some stats about the performance. 

```
// Old version first
// Total old: 1.420655265s, Total new: 131.683183ms, count 263, avg old: 5.401731ms, avg new: 500.696µs
// New version first
// Total old: 1.440066122s, Total new: 116.201786ms, count 253, avg old: 5.69196ms, avg new: 459.295µs
```
It doesn't matter a whole lot which handler was first, for serving `192` headers on an intel NUC with SSD, the current handler takes `~5.5ms`, the one in this PR takes about `0.5ms`, so one order of magnitude faster.