@holiman 2021-11-30
After letting it run for a bit longer: 
```
Total old: 7.335648596s, Total new: 709.078148ms, count 1208, avg old: 6.072556ms, avg new: 586.985µs
```